QVector<>::size() is an int, no need to cast it. I imagine this is a leftover from when std::vector<> was used (whose size is an unsigned type).
svn path=/trunk/KDE/kdeutils/kcalc/; revision=1171481

+++ stats.cpp
@@ -186,7 +186,7 @@ KNumber KStats::sample_std()
 
 int KStats::count() const
 {
-    return static_cast<int>(data_.size());
+    return data_.size();
 }
